requirements.txt
django
djangorestframework
psycopg2-binary
django-filter
python-dotenv==1.0.1

